+++ mlp_adsorption/gcmc.py
@@ -29,6 +29,7 @@ from mlp_adsorption.utilities import (
     random_position,
     random_rotation,
     vdw_overlap,
+    get_density
 )
 
 
@@ -120,7 +121,7 @@ class GCMC:
         self.set_framework(framework_atoms)
 
         # Get the framework density in g/cm^3
-        self.framework_density: float = self.get_framework_density()
+        self.framework_density: float = get_density(self.framework)
 
         # Adsorbate setup
         self.set_adsorbate(adsorbate_atoms)
@@ -238,16 +239,6 @@ class GCMC:
         self.adsorbate_energy = self.adsorbate.get_potential_energy()
         self.n_ads = len(self.adsorbate)
         self.adsorbate_mass = np.sum(self.adsorbate.get_masses()) / units.kg
-
-    def get_framework_density(self) -> float:
-        """
-        Get the density of the framework in g/cm^3
-        """
-
-        mass = np.sum(self.framework.get_masses()) / units.kg * 1e3    # Convert from amu to g
-        volume = self.V * (1e-8 ** 3)  # Convert from Angs^3 to cm^3
-
-        return mass / volume
 
     def set_state(self, state: ase.Atoms) -> None:
         """